+++ R/esqlabs-env.R
@@ -19,6 +19,10 @@ esqlabsEnv$widthPerPlotMapping <- 8
 # Default height of a plot of a single `PlotMapping`
 esqlabsEnv$heightPerPlotMapping <- 8
 
+#' Names of the settings stored in esqlabsEnv Can be used with `getEsqlabsRSetting()`
+#' @export
+esqlabsRSettingNames <- enum(names(esqlabsEnv))
+
 #' Get the value of a global esqlabsR setting.
 #'
 #' @param settingName String name of the setting
@@ -32,7 +36,7 @@ esqlabsEnv$heightPerPlotMapping <- 8
 #' getEsqlabsRSetting("packageName")
 getEsqlabsRSetting <- function(settingName) {
   if (!(any(names(esqlabsEnv) == settingName))) {
-    stop(messages$errorEsqlabsRSettingNotFound(settingName))
+    stop(messages$errorPackageSettingNotFound(settingName, esqlabsEnv))
   }
 
   obj <- esqlabsEnv[[settingName]]